+++ src-inst/flow_storm/remote_websocket_client.clj
@@ -55,7 +55,8 @@
 
                                      (onClose [code reason remote?]
                                        (log (format "Connection with %s closed. code=%s reson=%s remote?=%s"
-                                                    uri-str code reason remote?)))
+                                                    uri-str code reason remote?))
+                                       ((requiring-resolve 'flow-storm.api/stop)))
 
                                      (onError [^Exception e]
                                        (log-error (format "WebSocket error connection %s" uri-str) e)))]